feat: InspectionStatusTabs 컴포넌트 생성 (#202)
